// ui/src/dailogi-api/model/registerCommand.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * d-AI-logi API
 * API for d-AI-logi - an interactive web application for creating AI dialogues
 * OpenAPI spec version: v1.0.0
 */

/**
 * Command for user registration
 */
export interface RegisterCommand {
  /**
   * Username
   * @minLength 3
   * @maxLength 50
   */
  name: string;
  /**
   * Password
   * @minLength 6
   * @maxLength 2147483647
   */
  password: string;
  /**
   * Password confirmation
   * @minLength 1
   */
  passwordConfirmation: string;
}